# Change the behaviour of domain function with URLs containing UserInfo & @