--- a/src/Functions/URL/domainWithoutWWW.cpp
+++ b/src/Functions/URL/domainWithoutWWW.cpp
@@ -6,12 +6,16 @@ namespace DB
 {
 
 struct NameDomainWithoutWWW { static constexpr auto name = "domainWithoutWWW"; };
-using FunctionDomainWithoutWWW = FunctionStringToString<ExtractSubstringImpl<ExtractDomain<true>>, NameDomainWithoutWWW>;
+using FunctionDomainWithoutWWW = FunctionStringToString<ExtractSubstringImpl<ExtractDomain<true, false>>, NameDomainWithoutWWW>;
+
+struct NameDomainWithoutWWWRFC { static constexpr auto name = "domainWithoutWWWRFC"; };
+using FunctionDomainWithoutWWWRFC = FunctionStringToString<ExtractSubstringImpl<ExtractDomain<true, true>>, NameDomainWithoutWWWRFC>;
 
 
 REGISTER_FUNCTION(DomainWithoutWWW)
 {
     factory.registerFunction<FunctionDomainWithoutWWW>();
+    factory.registerFunction<FunctionDomainWithoutWWWRFC>();
 }
 
 }