--- a/src/Functions/URL/firstSignificantSubdomainCustom.cpp
+++ b/src/Functions/URL/firstSignificantSubdomainCustom.cpp
@@ -7,12 +7,15 @@ namespace DB
 {
 
 struct NameFirstSignificantSubdomainCustom { static constexpr auto name = "firstSignificantSubdomainCustom"; };
+using FunctionFirstSignificantSubdomainCustom = FunctionCutToFirstSignificantSubdomainCustomImpl<ExtractFirstSignificantSubdomain<true, false>, NameFirstSignificantSubdomainCustom>;
 
-using FunctionFirstSignificantSubdomainCustom = FunctionCutToFirstSignificantSubdomainCustomImpl<ExtractFirstSignificantSubdomain<true>, NameFirstSignificantSubdomainCustom>;
+struct NameFirstSignificantSubdomainCustomRFC { static constexpr auto name = "firstSignificantSubdomainCustomRFC"; };
+using FunctionFirstSignificantSubdomainCustomRFC = FunctionCutToFirstSignificantSubdomainCustomImpl<ExtractFirstSignificantSubdomain<true, true>, NameFirstSignificantSubdomainCustomRFC>;
 
 REGISTER_FUNCTION(FirstSignificantSubdomainCustom)
 {
     factory.registerFunction<FunctionFirstSignificantSubdomainCustom>();
+    factory.registerFunction<FunctionFirstSignificantSubdomainCustomRFC>();
 }
 
 }