--- a/src/Functions/URL/topLevelDomain.cpp
+++ b/src/Functions/URL/topLevelDomain.cpp
@@ -5,13 +5,18 @@
 namespace DB
 {
 
+template<bool conform_rfc>
 struct ExtractTopLevelDomain
 {
     static size_t getReserveLengthForElement() { return 5; }
 
     static void execute(Pos data, size_t size, Pos & res_data, size_t & res_size)
     {
-        std::string_view host = getURLHost(data, size);
+        std::string_view host;
+        if constexpr (conform_rfc)
+            host = getURLHostRFC(data, size);
+        else
+            host = getURLHost(data, size);
 
         res_data = data;
         res_size = 0;
@@ -41,11 +46,15 @@ struct ExtractTopLevelDomain
 };
 
 struct NameTopLevelDomain { static constexpr auto name = "topLevelDomain"; };
-using FunctionTopLevelDomain = FunctionStringToString<ExtractSubstringImpl<ExtractTopLevelDomain>, NameTopLevelDomain>;
+using FunctionTopLevelDomain = FunctionStringToString<ExtractSubstringImpl<ExtractTopLevelDomain<false>>, NameTopLevelDomain>;
+
+struct NameTopLevelDomainRFC { static constexpr auto name = "topLevelDomainRFC"; };
+using FunctionTopLevelDomainRFC = FunctionStringToString<ExtractSubstringImpl<ExtractTopLevelDomain<true>>, NameTopLevelDomainRFC>;
 
 REGISTER_FUNCTION(TopLevelDomain)
 {
     factory.registerFunction<FunctionTopLevelDomain>();
+    factory.registerFunction<FunctionTopLevelDomainRFC>();
 }
 
 }