--- a/tests/queries/0_stateless/00398_url_functions.reference
+++ b/tests/queries/0_stateless/00398_url_functions.reference
@@ -7,6 +7,14 @@ http
 
 ====HOST====
 www.example.com
+example.com
+example.com
+example.com
+example.com
+example.com
+example.com
+
+
 
 www.example.com
 127.0.0.1