--- a/tests/queries/0_stateless/02415_all_new_functions_must_be_documented.reference
+++ b/tests/queries/0_stateless/02415_all_new_functions_must_be_documented.reference
@@ -221,8 +221,12 @@ cutQueryString
 cutQueryStringAndFragment
 cutToFirstSignificantSubdomain
 cutToFirstSignificantSubdomainCustom
+cutToFirstSignificantSubdomainCustomRFC
 cutToFirstSignificantSubdomainCustomWithWWW
+cutToFirstSignificantSubdomainCustomWithWWWRFC
+cutToFirstSignificantSubdomainRFC
 cutToFirstSignificantSubdomainWithWWW
+cutToFirstSignificantSubdomainWithWWWRFC
 cutURLParameter
 cutWWW
 dateDiff
@@ -281,7 +285,9 @@ dictHas
 dictIsIn
 divide
 domain
+domainRFC
 domainWithoutWWW
+domainWithoutWWWRFC
 dotProduct
 dumpColumnStructure
 e
@@ -332,6 +338,8 @@ filesystemFree
 finalizeAggregation
 firstSignificantSubdomain
 firstSignificantSubdomainCustom
+firstSignificantSubdomainCustomRFC
+firstSignificantSubdomainRFC
 flattenTuple
 floor
 format
@@ -593,6 +601,7 @@ polygonsUnionSpherical
 polygonsWithinCartesian
 polygonsWithinSpherical
 port
+portRFC
 position
 positionCaseInsensitive
 positionCaseInsensitiveUTF8
@@ -898,6 +907,7 @@ toYearWeek
 today
 tokens
 topLevelDomain
+topLevelDomainRFC
 transactionID
 transactionLatestSnapshot
 transactionOldestSnapshot