… tests/queries/1_stateful/00038_uniq_state_merge2.reference