When UserInfo part and '@' appear in the URL, the host after @ should be returned. For example, when url is "https://user:pass@clickhouse.com/", start_of_host should be char 'c' after '@', end_of_host should be '/' other than ':'.


### Changelog category (leave one):

* New Feature



### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):

* Add `URL` Functions which conform rfc. Functions include: `cutToFirstSignificantSubdomainCustomRFC`, `cutToFirstSignificantSubdomainCustomWithWWWRFC`, `cutToFirstSignificantSubdomainRFC`, `cutToFirstSignificantSubdomainWithWWWRFC`, `domainRFC`, `domainWithoutWWWRFC`, `firstSignificantSubdomainCustomRFC`, `firstSignificantSubdomainRFC`, `portRFC`,  `topLevelDomainRFC`

For example, when url is "https://user:pass@clickhouse.com/", start_of_host should be char 'c' after '@', end_of_host should be '/' other than ':'.


Close https://github.com/ClickHouse/ClickHouse/issues/41262

> Information about CI checks: https://clickhouse.com/docs/en/development/continuous-integration/