@dentiscalprum 2022-10-14
> Some tests are failed because they use `domain` function for out `test.hits` dataset in stateful tests. Behaviour is changed, and it leads to a different result. We may check which rows were affected, validate if the current version is correct, and update these tests.

I found the address that solved fail and this situation has been adapted.