@dentiscalprum 2022-10-14
> Great! Could you also please some of them here?

This two address:
https://hurpass.com:3912%2F1.2&name=43&text=зайцев финансоножка&noreask
http://hurpass.com:3912&Sn=238699-wh-643-fw-1164427100

It seems that there are still addresses that cannot be domained, I will check again