@azat 2022-10-18
Also it will be great to show in changelog that it affects not only `domain` function everything that uses it internally:

```
src/Functions/URL/ExtractFirstSignificantSubdomain.h:        ExtractDomain<without_www>::execute(data, size, tmp, domain_length);
src/Functions/URL/ExtractFirstSignificantSubdomain.h:        ExtractDomain<without_www>::execute(data, size, tmp, domain_length);
src/Functions/URL/domain.cpp:using FunctionDomain = FunctionStringToString<ExtractSubstringImpl<ExtractDomain<false>>, NameDomain>;
src/Functions/URL/domain.h:inline std::string_view getURLHost(const char * data, size_t size)
src/Functions/URL/domain.h:struct ExtractDomain
src/Functions/URL/domain.h:        std::string_view host = getURLHost(data, size);
src/Functions/URL/domainWithoutWWW.cpp:using FunctionDomainWithoutWWW = FunctionStringToString<ExtractSubstringImpl<ExtractDomain<true>>, NameDomainWithoutWWW>;
src/Functions/URL/port.cpp:        std::string_view host = getURLHost(p, size);
src/Functions/URL/topLevelDomain.cpp:        std::string_view host = getURLHost(data, size);
src/Functions/URL/topLevelDomain.cpp:            /// since getURLHost() will not return a host if there is symbol after dot.
```

So:

- `port`
- `topLevelDomain`
- `domain`
- `cutToFirstSignificantSubdomain` (and all it's variants)