@azat 2022-10-18
>How does performance change?

Performance tests shows a little slowdown:



Old, s | New, s | Ratio of speedup (-) or slowdown (+) | Relative difference (new − old) / old | p < 0.01 threshold | Test | # | Query
-- | -- | -- | -- | -- | -- | -- | --
0.618 | 0.682 | +1.102x | 0.102 | 0.102 | url_hits | 1 | SELECT count() FROM hits_100m_single WHERE NOT ignore(domain(URL))

But only for raw domain, everything else is OK.
Maybe because now there are few more instructions for each domain, even though they can be omitted if you don't have userinfo in the URL.