@azat 2022-10-18
By the way, maybe it work marking this is backward incompatible change? Since someone may rely on the specific behavior of this function (for example if someone uses it for sharding).