@azat 2022-10-19
@dentiscalprum about changelog entry, I meant to use `Backward Incompatible Change` as `Changelog category` (https://raw.githubusercontent.com/ClickHouse/ClickHouse/master/.github/PULL_REQUEST_TEMPLATE.md).

Also AFAIR changelogs are not supports multiple list elements, so it is better to write it in one line somehow.