@azat 2022-10-20
>Maybe it will be convenient to keep the old functions under some names?

I would vote to introduce new functions with this new behavior not with old.