@alexey-milovidov 2022-10-20
@azat yes, it will be much better.

Let's name them:

domainPrecise
portPrecise
... etc.