@azat 2022-10-21
Maybe `domainRFC` and similar? (means RFC compatible)