@dentiscalprum 2022-10-23
> Maybe `domainRFC` and similar? (means RFC compatible)

@alexey-milovidov @azat @vdimir  Added RFC form of these functions, please review this PR