@azat 2022-10-25
Also worth to add performance tests for this new functions.